backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
redis==5.0.1
psycopg[binary]==3.3.3

sqlalchemy==2.0.25
alembic==1.13.1
prometheus-client==0.19.0
structlog==24.1.0
